Introduce a file size cache to minimize the need for file hashing
After some profiling, it seems like the majority of the time is spent in
hashing files. This commit introduces a file size cache that stores the
count of file sizes encountered as we traverse the directory tree for
files. The idea is that if a particular file's size is unique in the
directory tree, then there is no way it could be a duplicate and we can
avoid having to hash it.

On a test directory of 10,474 files with 1348 duplicates, runtime
performance was improved by 30% from 4 seconds to 2.7 seconds on
average.